moved db related functions to database.py
