+++ modules/data_processor.py
@@ -2,11 +2,8 @@ import os
 
 import pandas as pd
 from bs4 import BeautifulSoup
-from sqlalchemy.exc import OperationalError
-from sqlalchemy.orm import Session
 
 from modules.containers import detect_records
-from modules.database.database import JobOfferRecord, SessionLocal
 from modules.JobRecord import JobRecord
 from modules.websites import identify_website
 
@@ -47,66 +44,3 @@ def save_dataframe_to_csv(dataframe: pd.DataFrame, file_path: str) -> None:
     Save the given DataFrame to a CSV file.
     """
     dataframe.to_csv(file_path, index=False)
-
-
-def load_records_from_db(archive=False, all_records=False) -> pd.DataFrame:
-    """Load job records from the database."""
-    db: Session = SessionLocal()
-    try:
-        if not all_records:
-            if not archive:
-                records = db.query(JobOfferRecord).where(JobOfferRecord.offer_status == "active").all()
-            else:
-                records = db.query(JobOfferRecord).filter(JobOfferRecord.offer_status == "archived").all()
-        else:
-            records = db.query(JobOfferRecord).all()
-        data = [
-            {
-                "id": record.id,
-                "title": record.title,
-                "logo": record.logo,
-                "company_name": record.company_name,
-                "location": record.location,
-                "remote_status": record.remote_status,
-                "min_salary": record.min_salary,
-                "max_salary": record.max_salary,
-                "salary_details": record.salary_details,
-                "salary_text": record.salary_text,
-                "tags": record.tags,
-                "url": record.url,
-                "website": record.website,
-                "added_date": record.added_date,
-                "notes": record.notes,
-                "personal_rating": record.personal_rating,
-                "application_status": record.application_status,
-                "application_date": record.application_date,
-                "feedback_received": record.feedback_received,
-                "feedback_date": record.feedback_date,
-                "archived_date": record.archived_date,
-                "offer_status": record.offer_status,
-                "users_id": record.users_id,
-            }
-            for record in records
-        ]
-
-        return pd.DataFrame(data)
-    except OperationalError:
-        empty_frame = pd.DataFrame()
-        return empty_frame
-    finally:
-        db.close()
-
-
-def reactivate_all_offers():
-    """Reactivate all archived offers."""
-    db: Session = SessionLocal()
-    try:
-        records = db.query(JobOfferRecord).filter(JobOfferRecord.offer_status == "archived").all()
-        for record in records:
-            record.archived_date = None
-            record.offer_status = "active"
-        db.commit()
-
-        print(f"Reactivated {len(records)} offers.")
-    finally:
-        db.close()
