+++ pages/7_Archive_DB_Browser.py
@@ -2,8 +2,7 @@ import os
 
 import streamlit as st
 
-from modules.data_processor import load_records_from_db
-from modules.database.database import update_edited_dataframe
+from modules.database.database import load_records_from_db, update_edited_dataframe
 from modules.dataframe_settings import column_conversions
 
 st.set_page_config(layout="wide")
